requirements.txt
requests
requests[socks]
pytest-playwright